add nav component
