+++ webpack.config.js
@@ -13,6 +13,14 @@ module.exports = {
                 test: /\.css$/i,
                 use: ['style-loader','css-loader'],
             },
+            {
+                test: /\.(png|svg|jpg|jpeg|gif)$/i,
+                type: 'asset/resource',
+            },
+            {
+                test: /\.(woff|woff2|eot|ttf|otf)$/i,
+                type: 'asset/resource',
+            },
         ],
     },
 }
